Move database, depth choice
A file ~/.bcmoves may be used to avoid repeating mistakes:
it contains a record of board configurations, a count of games
they occured in, and the differences of winning and loosing
the game. This cured some of those really annoying, trivial
winning strategies against the AI.

A new algorithm for determining how many turns lookahead should
be used. Seems to work quite nice (by now it is even possible to
win against the simple levels, without being bored too much).